# docs/admin/swagger.yaml
basePath: /admin/warehouse
definitions:
...
  mgwarehouse.ConfigOrderConfirm:
    properties:
      confirmDelayInSeconds:
        type: integer
      isEnable:
        type: boolean
    type: object
...
  requestmodel.TimeRange:
    properties:
      from:
        type: integer
      to:
        type: integer
    type: object
...
  requestmodel.WarehouseCfgFood:
    properties:
      forceClosed:
        type: boolean
      timeRange:
        items:
          $ref: '#/definitions/requestmodel.TimeRange'
        type: array
    type: object
  requestmodel.WarehouseCfgFoodUpdate:
    properties:
      food:
        $ref: '#/definitions/requestmodel.WarehouseCfgFood'
    type: object
  requestmodel.WarehouseCfgOrder:
    properties:
      isLimitNumberOfPurchases:
        type: boolean
      limitNumberOfPurchases:
        type: integer
      minimumValue:
        type: number
      paymentMethod:
        $ref: '#/definitions/requestmodel.WarehouseCfgPaymentMethod'
    type: object
  requestmodel.WarehouseCfgOrderConfirm:
    properties:
      orderConfirm:
        $ref: '#/definitions/mgwarehouse.ConfigOrderConfirm'
    type: object
  requestmodel.WarehouseCfgOrderUpdate:
    properties:
      order:
        $ref: '#/definitions/requestmodel.WarehouseCfgOrder'
    type: object
  requestmodel.WarehouseCfgOther:
    properties:
      doesSupportSellyExpress:
        type: boolean
    type: object
  requestmodel.WarehouseCfgOtherUpdate:
    properties:
      other:
        $ref: '#/definitions/requestmodel.WarehouseCfgOther'
    type: object
  requestmodel.WarehouseCfgPartner:
    properties:
      authentication:
        type: string
      code:
        type: string
      enabled:
        type: boolean
      identityCode:
        type: string
    type: object
  requestmodel.WarehouseCfgPartnerUpdate:
    properties:
      partner:
        $ref: '#/definitions/requestmodel.WarehouseCfgPartner'
    type: object
  requestmodel.WarehouseCfgPaymentMethod:
    properties:
      bankTransfer:
        type: boolean
      cod:
        type: boolean
    type: object
  requestmodel.WarehouseCfgSupplier:
    properties:
      invoiceDeliveryMethod:
        type: string
    type: object
  requestmodel.WarehouseCfgSupplierUpdate:
    properties:
      supplier:
        $ref: '#/definitions/requestmodel.WarehouseCfgSupplier'
    type: object
...
  responsemodel.ResponseWarehouseConfigurationDetail:
    properties:
      autoConfirmOrder:
        $ref: '#/definitions/mgwarehouse.ConfigOrderConfirm'
      delivery:
        $ref: '#/definitions/responsemodel.ResponseWarehouseDeliveryConfig'
      food:
        $ref: '#/definitions/responsemodel.ResponseWarehousePartnerFood'
      id:
        type: string
      order:
        $ref: '#/definitions/responsemodel.ResponseWarehouseOrderConfig'
      other:
        $ref: '#/definitions/responsemodel.ResponseWarehouseOtherConfig'
      partner:
        $ref: '#/definitions/responsemodel.ResponseWarehousePartnerConfig'
      supplier:
        $ref: '#/definitions/responsemodel.ResponseWarehouseSupplierConfig'
      warehouse:
        type: string
    type: object
...
  responsemodel.ResponseWarehouseDeliveryConfig:
    properties:
      deliveryMethods:
        items:
          type: string
        type: array
      enabledSources:
        items:
          type: integer
        type: array
      priorityServiceCodes:
        items:
          type: string
        type: array
      types:
        items:
          type: string
        type: array
    type: object
...
  responsemodel.ResponseWarehouseOrderConfig:
    properties:
      isLimitNumberOfPurchases:
        type: boolean
      limitNumberOfPurchases:
        type: integer
      minimumValue:
        type: number
      paymentMethod:
        $ref: '#/definitions/responsemodel.ResponseWarehousePaymentMethodConfig'
    type: object
  responsemodel.ResponseWarehouseOtherConfig:
    properties:
      doesSupportSellyExpress:
        type: boolean
    type: object
  responsemodel.ResponseWarehousePartnerConfig:
    properties:
      authentication:
        type: string
      code:
        type: string
      enabled:
        type: boolean
      identityCode:
        type: string
    type: object
  responsemodel.ResponseWarehousePartnerFood:
    properties:
      forceClosed:
        type: boolean
      isClosed:
        type: boolean
      timeRange:
        items:
          $ref: '#/definitions/responsemodel.TimeRange'
        type: array
    type: object
  responsemodel.ResponseWarehousePaymentMethodConfig:
    properties:
      bankTransfer:
        type: boolean
      cod:
        type: boolean
    type: object
...
  responsemodel.ResponseWarehouseSupplierConfig:
    properties:
      invoiceDeliveryMethod:
        type: string
    type: object
...
  responsemodel.TimeRange:
    properties:
      from:
        type: integer
      to:
        type: integer
    type: object
  responsemodel.Upsert:
    properties:
      _id:
        type: string
    type: object
...
info:
  contact:
    email: dev@selly.vn
    name: Dev team
    url: https://selly.vn
  description: |
    All APIs for Warehouse admin.

    ******************************
    - Add description
    ******************************
  termsOfService: https://selly.vn
  title: Selly Warehouse - Admin API
  version: "1.0"
paths:
...
  /configurations/{id}:
    get:
      consumes:
      - application/json
      operationId: admin-warehouseConfiguration-detail
      parameters:
      - description: Device id
        in: header
        name: device-id
        required: true
        type: string
      - description: Warehouse id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodel.ResponseWarehouseConfigurationDetail'
      security:
      - ApiKeyAuth: []
      summary: Detail
      tags:
      - WarehouseConfiguration
...
  /configurations/{id}/food:
    put:
      consumes:
      - application/json
      operationId: admin-warehouseConfiguration-updateFood
      parameters:
      - description: Device id
        in: header
        name: device-id
        required: true
        type: string
      - description: Warehouse config id
        in: path
        name: id
        required: true
        type: string
      - description: Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requestmodel.WarehouseCfgFoodUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update Food
      tags:
      - WarehouseConfiguration
  /configurations/{id}/order:
    put:
      consumes:
      - application/json
      operationId: admin-warehouseConfiguration-updateOrder
      parameters:
      - description: Device id
        in: header
        name: device-id
        required: true
        type: string
      - description: Warehouse config id
        in: path
        name: id
        required: true
        type: string
      - description: Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requestmodel.WarehouseCfgOrderUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodel.Upsert'
      security:
      - ApiKeyAuth: []
      summary: Update
      tags:
      - WarehouseConfiguration
  /configurations/{id}/order-confirm:
    put:
      consumes:
      - application/json
      operationId: admin-warehouseConfiguration-order-confirm
      parameters:
      - description: Device id
        in: header
        name: device-id
        required: true
        type: string
      - description: Warehouse config id
        in: path
        name: id
        required: true
        type: string
      - description: Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requestmodel.WarehouseCfgOrderConfirm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: UpdateOrderConfirm
      tags:
      - WarehouseConfiguration
  /configurations/{id}/other:
    put:
      consumes:
      - application/json
      operationId: admin-warehouseConfiguration-updateOther
      parameters:
      - description: Device id
        in: header
        name: device-id
        required: true
        type: string
      - description: Warehouse config id
        in: path
        name: id
        required: true
        type: string
      - description: Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requestmodel.WarehouseCfgOtherUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodel.Upsert'
      security:
      - ApiKeyAuth: []
      summary: Update
      tags:
      - WarehouseConfiguration
  /configurations/{id}/partner:
    put:
      consumes:
      - application/json
      operationId: admin-warehouseConfiguration-updatePartner
      parameters:
      - description: Device id
        in: header
        name: device-id
        required: true
        type: string
      - description: Warehouse config id
        in: path
        name: id
        required: true
        type: string
      - description: Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requestmodel.WarehouseCfgPartnerUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodel.Upsert'
      security:
      - ApiKeyAuth: []
      summary: Update
      tags:
      - WarehouseConfiguration
  /configurations/{id}/supplier:
    put:
      consumes:
      - application/json
      operationId: admin-warehouseConfiguration-updateSupplier
      parameters:
      - description: Device id
        in: header
        name: device-id
        required: true
        type: string
      - description: Warehouse config id
        in: path
        name: id
        required: true
        type: string
      - description: Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requestmodel.WarehouseCfgSupplierUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodel.Upsert'
      security:
      - ApiKeyAuth: []
      summary: Update
      tags:
      - WarehouseConfiguration
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"